requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
aspose.slides>=23.0
spire.pdf>=8.6.0
langchain>=0.1.0
huggingface-hub>=0.19.0
transformers>=4.36.0
sentence-transformers==2.2.2
InstructorEmbedding>=1.0.1
faiss-cpu>=1.7.4
chromadb>=0.4.0
python-dotenv>=0.19.0
tiktoken>=0.5.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.65.0
nltk>=3.8.1
beautifulsoup4>=4.12.0